fix configure script to honor buildflags in Debian
When using the the buildflags which are exported when setting the
DPKG_EXPORT_BUILDFLAGS variable (or when trying to set some custom
buildflags manually) the build process failed with something like this:

  cc1: error: -Werror=format-security-g: no option -Wformat-security-g

This was caused by a missing whitespace when concatenating the CFLAGS
and LDFLAGS variables.

Thanks to Darshaka Pathirana <dpat@syn-net.org> for the patch!